+++ core/vil1/vil1_open.cxx
@@ -43,7 +43,7 @@ vil1_stream *vil1_open(char const* what, char const* how)
 
   if (!is) {
     // hacked check for filenames beginning "gen:".
-    int l = vcl_strlen(what);
+    int l = (int)vcl_strlen(what);
     if (l > 4 && vcl_strncmp(what, "gen:", 4) == 0) {
       if (vcl_strcmp(how, "r") == 0) {
         // Make an in-core stream...
@@ -65,7 +65,7 @@ vil1_stream *vil1_open(char const* what, char const* how)
 
   if (!is) {
     // maybe it's a URL?
-    int l = vcl_strlen(what);
+    int l = (int)vcl_strlen(what);
     if (l > 4 && vcl_strncmp(what, "http://", 7) == 0) {
 #ifdef __APPLE__
       vcl_cerr << __FILE__ ": cannot open URL for writing (yet)\n";
