+++ core/vsl/vsl_vector_io_bool.cxx
@@ -15,9 +15,9 @@ void vsl_b_write(vsl_b_ostream& s, const vcl_vector<bool>& v)
 {
   const short version_no = 1;
   vsl_b_write(s, version_no);
-  unsigned n = v.size();
+  unsigned int n = (unsigned int)(v.size());
   vsl_b_write(s,n);
-  for (unsigned i=0; i<n; ++i)
+  for (unsigned int i=0; i<n; ++i)
     vsl_b_write(s, v[i]);
 }
 
@@ -28,7 +28,7 @@ void vsl_b_read(vsl_b_istream& is, vcl_vector<bool>& v)
 {
   if (!is) return;
 
-  unsigned n;
+  unsigned int n;
   short ver;
   vsl_b_read(is, ver);
   switch (ver)
@@ -36,7 +36,7 @@ void vsl_b_read(vsl_b_istream& is, vcl_vector<bool>& v)
   case 1:
     vsl_b_read(is,n);
     v.resize(n);
-    for (unsigned i=0; i<n; ++i)
+    for (unsigned int i=0; i<n; ++i)
     {
       bool b;
       vsl_b_read(is, b);
